+++ joulescope/stream_buffer.pyx
@@ -45,7 +45,13 @@ DEF CAL_SAMPLE_SZ = 2 * 4  # sizeof(float)
 DEF STATS_FIELDS = 3  # current, voltage, power
 DEF STATS_VALUES = 4  # mean, variance, min, max
 DEF STATS_FLOATS_PER_SAMPLE = STATS_FIELDS * STATS_VALUES
-DEF SUPPRESS_SAMPLES_DEFAULT = 2
+DEF SUPPRESS_SAMPLES_DEFAULT = 3
+DEF I_RANGE_D_LENGTH = 3
+
+
+DEF SUPPRESS_MODE_OFF = 0  # disabled, zero delay
+DEF SUPPRESS_MODE_NORMAL = 1
+
 
 log = logging.getLogger(__name__)
 
@@ -377,9 +383,10 @@ cdef class StreamBuffer:
     cdef js_stream_buffer_reduction_s reductions[REDUCTION_MAX]
     cdef uint32_t reduction_count
 
+    cdef uint8_t i_range_d[I_RANGE_D_LENGTH]  # the i_range delay for processing
     cdef int32_t suppress_samples  # the total number of samples to suppress after range change
     cdef int32_t suppress_count  # the suppress counter, 1 = replace previous
-    cdef uint8_t i_range_last  # the last i_range for detection changes.
+    cdef uint8_t _suppress_mode
 
     cdef uint32_t stats_counter  # excludes NAN for mean
     cdef uint32_t stats_remaining
@@ -460,7 +467,10 @@ cdef class StreamBuffer:
 
         self.suppress_samples = SUPPRESS_SAMPLES_DEFAULT
         self.suppress_count = 0
-        self.i_range_last = 0xff
+        self._suppress_mode = SUPPRESS_MODE_NORMAL
+
+        for idx in range(I_RANGE_D_LENGTH):
+            self.i_range_d[idx] = 0xff
 
     def __init__(self, length, reductions):
         self.reset()
@@ -519,6 +529,26 @@ cdef class StreamBuffer:
     def voltage_range(self):
         return self.voltage_range
 
+    @property
+    def suppress_mode(self):
+        if self._suppress_mode == SUPPRESS_MODE_OFF:
+            return 'off'
+        elif self._suppress_mode == SUPPRESS_MODE_NORMAL:
+            return 'normal'
+        else:
+            return 'off'
+
+    @suppress_mode.setter
+    def suppress_mode(self, value):
+        if isinstance(value, str):
+            value = value.lower()
+        if value in [SUPPRESS_MODE_OFF, 'off', False] or value is None:
+            self._suppress_mode = SUPPRESS_MODE_OFF
+        elif value in [SUPPRESS_MODE_NORMAL, 'normal', 'on', True]:
+            self._suppress_mode = SUPPRESS_MODE_NORMAL
+        else:
+            raise ValueError('unsupported mode: %s' % (value, ))
+
     def status(self):
         return {
             'device_sample_id': {'value': self.device_sample_id, 'units': 'samples'},
@@ -565,7 +595,8 @@ cdef class StreamBuffer:
         self._energy_picojoules = 0
         self.stats_counter = 0
         self.suppress_count = 0
-        self.i_range_last = 0xff
+        for idx in range(I_RANGE_D_LENGTH):
+            self.i_range_d[idx] = 0xff
         for idx in range(REDUCTION_MAX):
             self.reductions[idx].sample_counter = 0
         self._stats_reset()
@@ -746,6 +777,7 @@ cdef class StreamBuffer:
         cdef uint32_t idx_start
         cdef uint32_t idx
         cdef uint32_t suppress_idx
+        cdef uint32_t i_range_idx
         cdef uint16_t raw_i
         cdef uint16_t raw_v
         cdef uint8_t i_range
@@ -764,10 +796,33 @@ cdef class StreamBuffer:
             idx = idx_start * 2
             raw_i = self.raw_ptr[idx + 0]
             raw_v = self.raw_ptr[idx + 1]
-            i_range = <uint8_t> ((raw_i & 0x0003) | ((raw_v & 0x0001) << 2))
-            if i_range != self.i_range_last and self.i_range_last < 255:
-                self.suppress_count = self.suppress_samples + 1
-            self.i_range_last = i_range
+            for i_range_idx in range(I_RANGE_D_LENGTH - 1, 0, -1):
+                self.i_range_d[i_range_idx] = self.i_range_d[i_range_idx - 1]
+            self.i_range_d[0] = <uint8_t> ((raw_i & 0x0003) | ((raw_v & 0x0001) << 2))
+
+            # process i_range for glitch suppression
+            if SUPPRESS_MODE_NORMAL == self._suppress_mode:
+                if self.i_range_d[0] == self.i_range_d[2]:
+                    # no change, use single delay since i_range leads data by 1 sample
+                    i_range = self.i_range_d[1]
+                elif self.i_range_d[0] == 0x7:
+                    # select is off or missing sample, use immediately
+                    i_range = self.i_range_d[0]
+                elif self.i_range_d[0] < self.i_range_d[2]:
+                    # use old select one more sample
+                    # moving to less sensitive range (smaller value resistor)
+                    i_range = self.i_range_d[2]
+                    if self.i_range_d[1] < self.i_range_d[2]:
+                        # delay suppress by one sample
+                        self.suppress_count = self.suppress_samples + 1
+                else:
+                    # Use new select immediately
+                    # moving to more sensitive range (larger value resistor)
+                    i_range = self.i_range_d[0]
+                    self.suppress_count = self.suppress_samples + 1
+            else:
+                i_range = self.i_range_d[0]
+
             sample_toggle_current = (raw_v >> 1) & 0x1
             raw_i = raw_i >> 2
             raw_v = raw_v >> 2
@@ -796,7 +851,8 @@ cdef class StreamBuffer:
 
             # Suppress Joulescope range switching glitch (at least for now).
             self.processed_sample_id += 1
-            if self.suppress_count > 0:
+
+            if self.suppress_count > 0:  # disable
                 if self.suppress_count == 1:
                     suppress_idx = idx_start + self.length - self.suppress_samples
                     while suppress_idx >= self.length: